准备区块转换
